First App Finished
* Implementation of StatefulWidget with it's characteristics, include
  manage of state for the roll dice random number and show the correctly
  dice image.
