+++ first_app/lib/gradient_container.dart
@@ -1,5 +1,5 @@
+import 'package:first_app/dice_roller.dart';
 import 'package:flutter/material.dart';
-import 'package:first_app/styled_text.dart';
 
 // Using variables inside Widget Classes
 
@@ -13,7 +13,7 @@ class GradientContainer extends StatelessWidget {
   // StatelessWidget needs `key` parameter
 
   // Here we add another parameter, the list of colors
-  GradientContainer({super.key, required this.colors});
+  const GradientContainer({super.key, required this.colors});
 
   // Another constructor, with defined colors
   GradientContainer.sunset({super.key}) : colors = [Colors.amber, Colors.red];
@@ -21,13 +21,6 @@ class GradientContainer extends StatelessWidget {
   // We add the final list of colors
   final List<Color> colors;
 
-  // Variable 
-  var activeDiceImage = 'assets/images/dice-2.png';
- 
-  void rollDice() {
-    // ...
-    activeDiceImage = 'assets/images/dice-4.png';
-  }
   // build method automatically receive by Flutter a values of context
   // context. Object that containts useful information
   @override
@@ -41,35 +34,9 @@ class GradientContainer extends StatelessWidget {
           end: endAlignment,
         ),
       ),
-      child: Center(
+      child: const Center(
         // Image Widget has a default constructor function called `asset`
-        child: Column(
-          // We need this mainAxisSize because Column and Row takes all the available space of container Widget
-          mainAxisSize: MainAxisSize.min,
-          children: [
-            Image.asset(
-              activeDiceImage,
-              width: 200,
-            ),
-            // SizedBox it's another way to put an space between Widgets
-            const SizedBox(
-              height: 20,
-            ),
-            TextButton(
-              onPressed: rollDice,
-              // We use styleFrom constructor for replace default style values of TextButton
-              style: TextButton.styleFrom(
-                // EdgeInsets allow to padding all or specific the padding position (top, left, right, bottom)
-                // padding: const EdgeInsets.only(
-                //   top: 20,
-                // ),
-                foregroundColor: Colors.white,
-                textStyle: const TextStyle(fontSize: 28),
-              ),
-              child: const Text("Roll Dice"),
-            )
-          ],
-        ),
+        child: DiceRoller(),
       ),
     );
   }
